refactor(preview): standardize scheduling rank
- Rename queue order fields to rank across frontend and Tauri boundaries.\n- Use u32 consistently for preview scheduling ranks.

+++ crates/oxy-domain/src/lib.rs
@@ -353,7 +353,7 @@ pub struct PreviewScheduleIntent {
     pub path: PathBuf,
     pub level: RenderLevel,
     pub priority: PreviewPriority,
-    pub rank: usize,
+    pub rank: u32,
 }
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
